+++ packages/preset-base/src/utilities/border.ts
@@ -1,5 +1,5 @@
 import type { UtilityConfig } from '@pandacss/types'
-import { createColorMixTransform } from './color-mix-transform'
+import { createColorMixTransform } from '../color-mix-transform'
 
 export const border: UtilityConfig = {
   borderRadius: {
